refactor: move node transfer logic to protocol
- This keeps the logic levels more consistent
- As client handling of transfers was already in protocol, it seems
more stringent to also keep the node handling of transfers there.
