+++ safenode/src/protocol/client_transfers/offline.rs
