+++ safenode/src/protocol/messages/event.rs
@@ -6,9 +6,9 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
-use crate::{
-    network_transfers::{Error, Result},
-    protocol::address::{dbc_address, DataAddress},
+use crate::protocol::{
+    address::{dbc_address, DataAddress},
+    node_transfers::{Error, Result},
 };
 
 use sn_dbc::SignedSpend;
