+++ safenode/src/protocol/node_transfers/error.rs
