+++ safenode/src/protocol/wallet/mod.rs
@@ -43,7 +43,7 @@ pub use self::{
     // network_store::NetworkWallet,
 };
 
-use super::transfers::{CreatedDbc, Outputs as TransferDetails};
+use super::client_transfers::{CreatedDbc, Outputs as TransferDetails};
 
 use sn_dbc::{Dbc, DbcIdSource, DerivedKey, PublicAddress, Token};
